implemented sync to firebase for tasks

+++ app/src/main/java/com/example/deltatlog/ui/ProjectFragment.kt
@@ -19,6 +19,7 @@ import androidx.room.withTransaction
 import com.example.deltatlog.R
 import com.example.deltatlog.data.datamodels.Project
 import com.example.deltatlog.data.local.getDatabase
+import com.example.deltatlog.data.local.getTaskDatabase
 import com.example.deltatlog.databinding.FragmentProjectBinding
 import com.example.deltatlog.viewModel
 import com.google.firebase.auth.FirebaseAuth
@@ -63,138 +64,59 @@ class ProjectFragment : Fragment() {
         val currentUserId = firebaseAuth.currentUser!!.uid
 
 
-
-            // Set onClickListener on menu item logout
-            binding.materialToolbar.setOnMenuItemClickListener {
-                when (it.itemId) {
-                    R.id.logout ->
-                        firebaseAuth.signOut()
-                }
-                findNavController().navigate(ProjectFragmentDirections.actionHomeFragmentToLoginFragment())
-                if (firebaseAuth.currentUser == null) {
-                    Toast.makeText(
-                        context,
-                        "Successfully logged out user $currentUserEmail",
-                        Toast.LENGTH_LONG
-                    ).show()
-                }
-                true
+        // Set onClickListener on menu item logout
+        binding.materialToolbar.setOnMenuItemClickListener {
+            when (it.itemId) {
+                R.id.logout ->
+                    firebaseAuth.signOut()
             }
+            findNavController().navigate(ProjectFragmentDirections.actionHomeFragmentToLoginFragment())
+            if (firebaseAuth.currentUser == null) {
+                Toast.makeText(
+                    context,
+                    "Successfully logged out user $currentUserEmail",
+                    Toast.LENGTH_LONG
+                ).show()
+            }
+            true
+        }
 
-            val recyclerView = binding.projectList
+        val recyclerView = binding.projectList
 
-            viewModel.projectList.observe(
-                viewLifecycleOwner,
-                Observer {
-                    recyclerView.adapter =
-                        ProjectAdapter(viewModel, requireContext(), it, lifecycleScope)
-                }
-            )
-            binding.floatingActionButton.setOnClickListener {
+        viewModel.projectList.observe(
+            viewLifecycleOwner,
+            Observer {
+                recyclerView.adapter =
+                    ProjectAdapter(viewModel, requireContext(), it, lifecycleScope)
+            }
+        )
+        binding.floatingActionButton.setOnClickListener {
 
-                val builder = AlertDialog.Builder(context)
-                val inflater = layoutInflater
-                val dialogLayout = inflater.inflate(R.layout.edit_text_dialogue_project, null)
-                val newProjectName = dialogLayout.findViewById<EditText>(R.id.input_project_name)
-                val newCustomerName =
-                    dialogLayout.findViewById<EditText>(R.id.input_project_customer_name)
-                val newDescription =
-                    dialogLayout.findViewById<EditText>(R.id.input_project_description)
-                val newCompanyName = dialogLayout.findViewById<EditText>(R.id.input_company_name)
+            val builder = AlertDialog.Builder(context)
+            val inflater = layoutInflater
+            val dialogLayout = inflater.inflate(R.layout.edit_text_dialogue_project, null)
+            val newProjectName = dialogLayout.findViewById<EditText>(R.id.input_project_name)
+            val newCustomerName =
+                dialogLayout.findViewById<EditText>(R.id.input_project_customer_name)
+            val newDescription =
+                dialogLayout.findViewById<EditText>(R.id.input_project_description)
+            val newCompanyName = dialogLayout.findViewById<EditText>(R.id.input_company_name)
 
-                with(builder) {
-                    setTitle("New Project")
-                    setPositiveButton("Ok") { dialog, which ->
-                        val newProjectNameString = newProjectName.text.toString()
-                        val newCustomerNameString = newCustomerName.text.toString()
-                        val newDescriptionString = newDescription.text.toString()
-                        val newCompanyNameString = newCompanyName.text.toString()
-                        val newProject = Project()
+            with(builder) {
+                setTitle("New Project")
+                setPositiveButton("Ok") { dialog, which ->
+                    val newProjectNameString = newProjectName.text.toString()
+                    val newCustomerNameString = newCustomerName.text.toString()
+                    val newDescriptionString = newDescription.text.toString()
+                    val newCompanyNameString = newCompanyName.text.toString()
+                    val newProject = Project()
 
-                        if (newCompanyNameString != "") {
-                            viewModel.loadLogo(newCompanyNameString) {
-                                Log.d("ProjectFragment", "(5) Here updating logourl")
-                                Log.d("ProjectFragment", viewModel.logoLiveData.value!!.logo)
+                    if (newCompanyNameString != "") {
+                        viewModel.loadLogo(newCompanyNameString) {
+                            Log.d("ProjectFragment", "(5) Here updating logourl")
+                            Log.d("ProjectFragment", viewModel.logoLiveData.value!!.logo)
 
-                                newProject.logoUrl = viewModel.logoLiveData.value!!.logo
-
-                                if (newProjectNameString != "") {
-                                    newProject.name = newProjectNameString
-                                }
-                                if (newCustomerNameString != "") {
-                                    newProject.nameCustomer = newCustomerNameString
-                                }
-                                if (newDescriptionString != "") {
-                                    newProject.description = newDescriptionString
-                                }
-                                if (newCompanyNameString != "") {
-                                    newProject.companyName = newCompanyNameString
-                                }
-                                Log.d("ProjectFragment", newProject.logoUrl)
-
-                                viewModel.insertProject(newProject) {
-
-                                    val database = getDatabase(context)
-
-                                    // Launch a coroutine
-                                    lifecycleScope.launch {
-                                        // Perform the database operation within the coroutine
-                                        val projectList: List<Project> =
-                                            withContext(Dispatchers.IO) {
-                                                database.projectDatabaseDao.getAllNLD()
-                                            }
-
-                                        val project2Add = projectList.last()
-
-                                        Log.d("ProjectFragment", currentUserEmail!!)
-                                        Log.d("ProjectFragment", currentUserId)
-
-                                        val firebaseItem2Add = hashMapOf(
-                                            "id" to project2Add.id,
-                                            "name" to project2Add.name,
-                                            "date" to project2Add.date,
-                                            "nameCustomer" to project2Add.nameCustomer,
-                                            "companyName" to project2Add.companyName,
-                                            "homepage" to project2Add.homepage,
-                                            "logoUrl" to project2Add.logoUrl,
-                                            "image" to project2Add.image,
-                                            "date" to project2Add.date,
-                                            "description" to project2Add.description,
-                                            "color" to project2Add.color,
-                                            "numberOfTasks" to project2Add.numberOfTasks,
-                                            "totalTime" to project2Add.totalTime
-                                        )
-
-                                        Log.d("firebasePID", project2Add.id.toString())
-
-                                        db.collection("users").document(currentUserId)
-                                            .collection("projects")
-                                            .document(project2Add.id.toString())
-                                            .set(firebaseItem2Add)
-                                            .addOnSuccessListener {
-                                                Log.d(
-                                                    "firebase",
-                                                    "DocumentSnapshot successfully written!"
-                                                )
-                                            }
-                                            .addOnFailureListener { e ->
-                                                Log.w(
-                                                    "firebase",
-                                                    "Error writing document",
-                                                    e
-                                                )
-                                            }
-
-                                        Toast.makeText(
-                                            context,
-                                            "$newProjectNameString created",
-                                            Toast.LENGTH_SHORT
-                                        )
-                                            .show()
-                                    }
-                                }
-                            }
-                        } else {
+                            newProject.logoUrl = viewModel.logoLiveData.value!!.logo
 
                             if (newProjectNameString != "") {
                                 newProject.name = newProjectNameString
@@ -217,13 +139,14 @@ class ProjectFragment : Fragment() {
                                 // Launch a coroutine
                                 lifecycleScope.launch {
                                     // Perform the database operation within the coroutine
-                                    val projectList: List<Project> = withContext(Dispatchers.IO) {
-                                        database.projectDatabaseDao.getAllNLD()
-                                    }
+                                    val projectList: List<Project> =
+                                        withContext(Dispatchers.IO) {
+                                            database.projectDatabaseDao.getAllNLD()
+                                        }
 
                                     val project2Add = projectList.last()
 
-                                    Log.d("ProjectFragment", project2Add.id.toString())
+                                    Log.d("ProjectFragment", currentUserEmail!!)
                                     Log.d("ProjectFragment", currentUserId)
 
                                     val firebaseItem2Add = hashMapOf(
@@ -241,7 +164,9 @@ class ProjectFragment : Fragment() {
                                         "numberOfTasks" to project2Add.numberOfTasks,
                                         "totalTime" to project2Add.totalTime
                                     )
+
                                     Log.d("firebasePID", project2Add.id.toString())
+
                                     db.collection("users").document(currentUserId)
                                         .collection("projects")
                                         .document(project2Add.id.toString())
@@ -259,6 +184,7 @@ class ProjectFragment : Fragment() {
                                                 e
                                             )
                                         }
+
                                     Toast.makeText(
                                         context,
                                         "$newProjectNameString created",
@@ -268,20 +194,94 @@ class ProjectFragment : Fragment() {
                                 }
                             }
                         }
+                    } else {
+
+                        if (newProjectNameString != "") {
+                            newProject.name = newProjectNameString
+                        }
+                        if (newCustomerNameString != "") {
+                            newProject.nameCustomer = newCustomerNameString
+                        }
+                        if (newDescriptionString != "") {
+                            newProject.description = newDescriptionString
+                        }
+                        if (newCompanyNameString != "") {
+                            newProject.companyName = newCompanyNameString
+                        }
+                        Log.d("ProjectFragment", newProject.logoUrl)
+
+                        viewModel.insertProject(newProject) {
+
+                            val database = getDatabase(context)
+
+                            // Launch a coroutine
+                            lifecycleScope.launch {
+                                // Perform the database operation within the coroutine
+                                val projectList: List<Project> = withContext(Dispatchers.IO) {
+                                    database.projectDatabaseDao.getAllNLD()
+                                }
+
+                                val project2Add = projectList.last()
+
+                                Log.d("ProjectFragment", project2Add.id.toString())
+                                Log.d("ProjectFragment", currentUserId)
+
+                                val firebaseItem2Add = hashMapOf(
+                                    "id" to project2Add.id,
+                                    "name" to project2Add.name,
+                                    "date" to project2Add.date,
+                                    "nameCustomer" to project2Add.nameCustomer,
+                                    "companyName" to project2Add.companyName,
+                                    "homepage" to project2Add.homepage,
+                                    "logoUrl" to project2Add.logoUrl,
+                                    "image" to project2Add.image,
+                                    "date" to project2Add.date,
+                                    "description" to project2Add.description,
+                                    "color" to project2Add.color,
+                                    "numberOfTasks" to project2Add.numberOfTasks,
+                                    "totalTime" to project2Add.totalTime
+                                )
+                                Log.d("firebasePID", project2Add.id.toString())
+                                db.collection("users").document(currentUserId)
+                                    .collection("projects")
+                                    .document(project2Add.id.toString())
+                                    .set(firebaseItem2Add)
+                                    .addOnSuccessListener {
+                                        Log.d(
+                                            "firebase",
+                                            "DocumentSnapshot successfully written!"
+                                        )
+                                    }
+                                    .addOnFailureListener { e ->
+                                        Log.w(
+                                            "firebase",
+                                            "Error writing document",
+                                            e
+                                        )
+                                    }
+                                Toast.makeText(
+                                    context,
+                                    "$newProjectNameString created",
+                                    Toast.LENGTH_SHORT
+                                )
+                                    .show()
+                            }
+                        }
                     }
-                    setNegativeButton("Cancel") { dialog, which ->
-                        dialog.dismiss()
-                        Toast.makeText(
-                            context,
-                            "Project creation cancelled by user",
-                            Toast.LENGTH_SHORT
-                        ).show()
-                    }
-                    setView(dialogLayout)
-                    show()
                 }
+                setNegativeButton("Cancel") { dialog, which ->
+                    dialog.dismiss()
+                    Toast.makeText(
+                        context,
+                        "Project creation cancelled by user",
+                        Toast.LENGTH_SHORT
+                    ).show()
+                }
+                setView(dialogLayout)
+                show()
             }
-            // here firebase stuff
+        }
+        // here firebase stuff
         val db = Firebase.firestore
         val projectCollection = db.collection("users").document(currentUserId)
             .collection("projects")
@@ -289,8 +289,14 @@ class ProjectFragment : Fragment() {
         // TODO Snapshot listener for firebase changes
 
         val database = getDatabase(requireContext())
-        lifecycleScope.launch {
-            database.projectDatabaseDao.deleteAllProjects()
+        val taskDatabase = getTaskDatabase(requireContext())
+
+        if (viewModel.databaseDeleted == false) {
+            lifecycleScope.launch {
+                database.projectDatabaseDao.deleteAllProjects()
+                taskDatabase.taskDatabaseDao.deleteAllTasks()
+            }
+            viewModel.databaseDeleted = true
         }
 
         projectCollection.addSnapshotListener { snapshot, exception ->
@@ -350,7 +356,6 @@ class ProjectFragment : Fragment() {
                     database.projectDatabaseDao.insertAll(projectsToInsert)
                 }
             }
-
         }
     }
 }
